Fixed issue #13

+++ auth.go
@@ -47,7 +47,7 @@ func (a *Auth) LogOn(details LogOnDetails) {
 	logon.ProtocolVersion = proto.Uint32(MsgClientLogon_CurrentProtocol)
 	logon.ShaSentryfile = details.SentryFileHash
 
-	atomic.StoreUint64(&a.client.steamId, uint64(NewIdAdv(0, 1, int32(EUniverse_Public), EAccountType_Individual)))
+	atomic.StoreUint64(&a.client.steamId, uint64(NewIdAdv(0, 1, int32(EUniverse_Public), int32(EAccountType_Individual))))
 
 	a.client.Write(NewClientMsgProtobuf(EMsg_ClientLogon, logon))
 }
